datatype QName reworked for saving space and time
QName is not longer a record of 3 components for
prefix, local part and namespace uri, but a data type
with 3 constructors for simple names (LP), prefix/local part (PX)
and an additional namespace uri (NS). The representation of
a name is always unique, a simple name is just an (LP name) value,
a px:lp name is (PX px (LP lp)) and a namespace is attached around
a name (NS uri (LP lp)) for default namespaces and (NS uri (PX px (LP lp)))
for expicit namespace prefixes. Further the strings for names are collected in a
global map. When constructing a name this map is looked up for the string,
and the already cached version of the string is used for building up the name.
This leads to sharing of names in XML documents an saves space.

darcs-hash:20081124175458-be05e-d6bcf72e5cb5a48ba88be4c7426c8179f36ebe91.gz